requirement.txt
streamlit==1.28.1
pandas==2.1.1
numpy==1.26.4
joblib==1.3.2
Pillow==10.1.0
scikit-learn==1.3.2
